+++ config/schedule.rb
@@ -4,3 +4,8 @@
 every 1.hour do
   runner 'Import::SyncCompaniesJob.perform_later'
 end
+
+# Send insights notifications
+every 4.hours do
+  runner 'Export::SendInsightsNotificationJob.perform_later'
+end