+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- configuring Slack webhook url for notifications
+- sending insights notifications to Slack
+
 ## [1.1.6] - 2023-01-25
 ### Modified
 - speed improvements for importing pull requests